Task2.V3

+++ Tyuiu.MoskalevEA.Sprint1.Task2.V3/Program.cs
@@ -38,7 +38,7 @@ namespace Tyuiu.MoskalevEA.Sprint1.Task2.V3
             Console.WriteLine("* РЕЗУЛЬТАТ:                                                             *");
             Console.WriteLine("**************************************************************************");
 
-            Console.WriteLine("Минут в часах" + ds.ConvertHourToMin(x));
+            Console.WriteLine("Минут в часах" + ds.ConvertHourToMin((int)x));
            
             Console.ReadLine();
         }
